pyproject.toml
[project]
name = "SuddAI"
version = "0.1.0"
description = ""
authors = [
    { name = "Wuor Bhang", email = "uhuribhang211@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "folium>=0.20.0",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "pillow>=11.3.0",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "streamlit-folium>=0.25.0",
]
